Complete navigation switch also for the content (previously only AppBar was done)
- Remove intermediate unneded components
- Clean old imports and sort CSS by specifity
