--- apps/projects/app/components/Content/IssueDetail/components/index.js
@@ -1,2 +1,0 @@
-export { default as Detail } from './Detail'
-export { default as Navigation } from './Navigation'
