+++ apps/projects/app/components/Content/Issues.js
@@ -178,7 +178,7 @@ class Issues extends React.PureComponent {
       // if we look for all funded issues, regardless of stage...
       let filterPass =
         status in filters.statuses ||
-        ('all-funded' in filters.statuses && status !== 'not-funded')
+          ('all-funded' in filters.statuses && status !== 'not-funded')
           ? true
           : false
       // ...or at specific stages
@@ -215,10 +215,13 @@ class Issues extends React.PureComponent {
   }
 
   handleIssueClick = issue => {
+    console.log('issue clicked')
+
     this.props.setIssueDetail(true)
     this.setState({ currentIssue: issue })
   }
 
+  // TODO: Not used, left here for reference (if we need to reset currentIssue)
   handleIssueDetailClose = () => {
     this.props.setIssueDetail(false)
     this.setState({ currentIssue: null })
@@ -453,7 +456,6 @@ class Issues extends React.PureComponent {
     return (
       <IssueDetail
         issue={currentIssueShaped}
-        onClose={this.handleIssueDetailClose}
         onReviewApplication={onReviewApplication}
         onRequestAssignment={onRequestAssignment}
         onSubmitWork={onSubmitWork}
@@ -521,7 +523,7 @@ class Issues extends React.PureComponent {
       onReviewApplication,
       onSubmitWork,
       onReviewWork,
-      showIssueDetail
+      issueDetail
     } = this.props
 
     const { currentIssue, filters } = this.state
@@ -530,7 +532,7 @@ class Issues extends React.PureComponent {
     if (projects.length === 0) return <Empty action={onNewProject} />
 
     // same if we only need to show Issue's Details screen
-    if (showIssueDetail)
+    if (issueDetail)
       return this.renderCurrentIssue(currentIssue, this.props)
 
     const currentSorter = this.generateSorter()
