+++ apps/projects/app/index.html
@@ -5,7 +5,7 @@
   </head>
 
   <body>
-    <div id="projects"></div>
+    <div id="projects" style="overflow: hidden"></div>
     <script src="index.js"></script>
   </body>
 </html>
